+++ scripts/generate_noise_images.py
@@ -5,7 +5,7 @@ import torch
 import time
 from tqdm import tqdm
 
-from stable_diffusion.util import save_images
+from stable_diffusion.utils.model import save_images
 from cli_builder import CLI
 
 noise_seeds = [
