+++ stable_diffusion/utils/cli.py
