+++ stable_diffusion/utils/model.py
@@ -8,7 +8,6 @@ summary: >
 # Utility functions for [stable diffusion](index.html)
 """
 
-import os
 import random
 from pathlib import Path
 from typing import Union
